Remove unsupported syscall event for setsockopt(*, SOL_SOCKET, SO_OOBINLINE).
Netstack behaves as if SO_OOBINLINE is always set, and was logging an
unsupported syscall event if the app tries to disable it. We don't have a real
use case for TCP urgent mechanisms (and RFC6093 says apps SHOULD NOT use it).

This CL keeps the current behavior, but removes the unsupported syscall event.

Fixes #6123

PiperOrigin-RevId: 377553466